requirements.txt
pynput
pystray
pillow
tk
